requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-groq==0.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.0
pylint==3.0.3
black==23.12.1
mypy==1.8.0
